Fixed compilation errors and some warnings

+++ gforce.F90
@@ -3,5 +3,5 @@ module gforce
   implicit none
   save
 
-  real :: dSdpi(ksizex_l, ksizey_l, ksizet_l, 3)
+  real(dp)  :: dSdpi(ksizex_l, ksizey_l, ksizet_l, 3)
 end module
